Port more emails to notification templates
