--- app/views/events_mailer/event_updated.html.haml
@@ -1,11 +1,0 @@
-%h1 Event updated
-
-%p
-  "#{@event.title}" has been updated.  To view the event, go
-  #{link_to 'here', event_url_for_convention(@event)}.
-
-%hr
-
-%h2 Change log
-
-= FormResponseChangeGroupPresenter.new(@changes, @event.convention).html
