--- app/views/events_mailer/event_updated.text.erb
@@ -1,7 +1,0 @@
-"<%= @event.title %>" has been updated.
-To view the event, go to <%= event_url_for_convention(@event) %>.
-
-------
-
-A changelog is available; please view this email in an HTML-enabled mail
-client to see it.
