--- app/views/user_activity_alert_mailer/alert.html.haml
@@ -1,12 +1,0 @@
-%h1 Alert
-
-%p
-  #{@alert_user_con_profile.name} (#{@alert_user_con_profile.email}) #{event_description(@event, @alert_user_con_profile)}
-  for #{@alert_user_con_profile.convention.name}.
-
-%p
-  This alert was triggered because:
-
-%ul
-  - trigger_reasons(@user_activity_alert, @alert_user_con_profile).each do |reason|
-    %li= reason
