--- app/views/user_activity_alert_mailer/alert.text.erb
@@ -1,11 +1,0 @@
-Alert
-=====
-
-<%= @alert_user_con_profile.name %> (<%= @alert_user_con_profile.email %>)
-<%= event_description(@event, @alert_user_con_profile) %> for <%= @alert_user_con_profile.convention.name %>.
-
-This alert was triggered because:
-
-<% trigger_reasons(@user_activity_alert, @alert_user_con_profile).each do |reason| -%>
-* <%= reason %>
-<% end -%>
